requirements.txt
asgiref==3.2.10
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.1
entrypoints==0.3
Flask==1.1.2
Flask-MySQL==1.5.1
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Pygments==2.6.1
PyMySQL==0.10.0
pyrsistent==0.16.0
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
Send2Trash==1.5.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
